feat: add melpazoid build checks to GitHub workflows
This commit adds a new file named melpazoid.yml to the .github/workflows directory. This file enables melpazoid's checks for the package and is configured to work with GitHub actions. It also includes setup steps for Python 3.10 and the installation of the required package dependencies. The job will run on push and pull request events.ci