feat: add file extension to Org mode language mapping
This commit introduces a new customization option
`gptel-extra-ext-to-org-langs-alist` which maps file extensions to
their corresponding Org mode languages. This allows for better
integration with Org mode when dealing with various programming
languages and file types.

Additionally, a new default save directory `gptel-extra-default-save-dir`
is defined for saving GPT-EL extra state files. This directory is
resolved relative to the user's Emacs directory and will be created
when saving a state file for the first time.

Several new functions are added to enhance the functionality of the
package:
- `gptel-extra-trim-org-src-markdown-block` trims Markdown source
  blocks in Org content.
- `gptel-extra--include-files` expands included files in Org content.
- `gptel-extra-denormalize-prompt-item` modifies a property list based
  on the role and content trimming.
- `gptel-extra-normalize-prompts` produces a list of arguments for
  calling Curl with normalized prompts.
- `gptel-extra-filter-curl--get-args` normalizes prompt in arguments
  for Curl call.
- `gptel-extra-get-files`, `gptel-extra-copy-dir-tree-as-org-block`,
  `gptel-extra-magit-staged-files`, `gptel-extra-get-dired-marked-files`,
  `gptel-extra-copy-staged-files-contents-as-org-blocks`,
  `gptel-extra-copy-files-contents-as-org-blocks`,
  `gptel-extra-get-file-content`, `gptel-extra-get-project-relative-name`,
  `gptel-extra-set-default-dir-to-project-dir`, `gptel-extra-save`,
  `gptel-extract-saved`, `gptel-extra-file-content-as-org-block`, and
  `gptel-extra-copy-file-contents-as-org-blocks` are utility functions
  that provide various features such as copying directory trees as Org
  blocks, handling staged Git files, saving buffer content with a
  timestamp, and more.

The `gptel-extra-mode` minor mode is updated to use the new
`gptel-extra-filter-curl--get-args` advice.

This enhancement improves the usability of the GPT-EL package by
providing additional tools and options for users to work with files and
Org mode more effectively.